requirements.txt
Django==4.2.21
celery==5.4.0
cryptography==44.0.3
django-filter==25.1
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
flower==2.0.1
# grpcio==1.67.0
# grpcio-status==1.67.0
# grpcio-tools==1.67.0
msgpack==1.1.0
psycopg-binary==3.2.3
django-cors-headers==4.7.0
webauthn==2.5.2
firebase-admin==6.6.0
twilio==9.5.0
phonenumbers==8.13.44
user-agents==2.2.0
djangorestframework-simplejwt==5.3.1